requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
amplitude-analytics==1.1.4
pytz==2024.1
